# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/postagens/{id}/imagem": {
            "post": {
                "description": "Recebe um arquivo de imagem e o UUID de uma postagem. Retorna status 202 em caso de sucesso.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "postagens"
                ],
                "summary": "Upload de imagem de postagem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID da postagem",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Arquivo de imagem (máx 10MB)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Status Accepted (202)"
                    },
                    "400": {
                        "description": "Erro na leitura da imagem ou ID inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno no upload",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deleta uma imagem do S3 e remove a URL do campo imagem_url da tabela postagens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "postagens"
                ],
                "summary": "Remove uma imagem da postagem",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Postagem",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "URL da imagem a ser removida",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/imagenshandler.DeletarImagemRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Imagem deletada com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                }
            }
        },
...
        "/trocas/{id}/imagem": {
            "post": {
                "description": "Recebe um arquivo de imagem e o UUID de uma troca. Retorna status 202 em caso de sucesso.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trocas"
                ],
                "summary": "Upload de imagem de troca",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID da troca",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Arquivo de imagem (máx 10MB)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Status Accepted (202)"
                    },
                    "400": {
                        "description": "Erro na leitura da imagem ou ID inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno no upload",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "imagenshandler.DeletarImagemRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}